fix(ps-cloud): make guest health checks more resilient SIT-372
Currently if the system never runs the health check the script will
always return unknown meaning the node is never flagged as unhealthy.

Additionally if the system goes readonly then the last output of the
guest health will be read in perpuitity. We want to make sure we are
only getting fresh data.